imputation/pyproject.toml
[project]
name = "imputation"
version = "0.1.0"
description = ""
authors = [
    {name = "Tylores",email = "tylor.slay@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <3.12"
dependencies = [
"opendssdirect-py>=0.7.0",
"absl-py==2.2.2",
"astunparse==1.6.3",
"certifi==2025.4.26",
"charset-normalizer==3.4.1",
"contourpy==1.3.2",
"cycler==0.12.1",
"flatbuffers==25.2.10",
"fonttools==4.57.0",
"gast==0.6.0",
"google-pasta==0.2.0",
"grpcio==1.71.0",
"h5py==3.13.0",
"idna==3.10",
"joblib==1.4.2",
"keras==3.9.2",
"kiwisolver==1.4.8",
"libclang==18.1.1",
"Markdown==3.8",
"markdown-it-py==3.0.0",
"MarkupSafe==3.0.2",
"matplotlib==3.10.1",
"mdurl==0.1.2",
"ml_dtypes==0.5.1",
"namex==0.0.9",
"narwhals==1.44.0",
"numpy==2.1.3",
"opt_einsum==3.4.0",
"optree==0.15.0",
"packaging==25.0",
"pandas==2.2.3",
"pillow==11.2.1",
"plotly==6.2.0",
"protobuf==5.29.4",
"Pygments==2.19.1",
"pyparsing==3.2.3",
"python-dateutil==2.9.0.post0",
"pytz==2025.2",
"requests==2.32.3",
"rich==14.0.0",
"scikit-learn==1.6.1",
"scipy==1.15.2",
"seaborn==0.13.2",
"six==1.17.0",
"tensorboard==2.19.0",
"tensorboard-data-server==0.7.2",
"tensorflow==2.19.0",
"tensorflow-io-gcs-filesystem==0.31.0",
"termcolor==3.0.1",
"threadpoolctl==3.6.0",
"typing_extensions==4.13.2",
"tzdata==2025.2",
"urllib3==2.4.0",
"Werkzeug==3.1.3",
"wrapt==1.17.2",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
